test refactor: example
